[WIP] Add initial samplex serialization support
This commit adds a first pass for a serialization of the Samplex object,
it leverages the built-in node-link JSON format serializer built in to
rustworkx for a fast JSON generation. While JSON is a bit more verbose
than a binary representations, the advantage of this is that it's very
simple to implement using the infrastructure built in to rustworkx.
Additionally, node-link json is based on d3.js's object for graph
visualization. This means that there is existing tooling out there that
can consume and generate this format. One example, might be d3 based
visualization in IQP's job viewer. This also doesn't preclude other
serialization formats in the future, this is just a starting point to
get something going quickly.

TODO:
- [ ] Add node deserialization functions
- [ ] Expand test coverage