Improve `PrePropagate` merging (#23)
## Summary
This PR improves the performance of the `PrePropagate` merging, which
takes a considerable time of the building time. Currently every new node
is compared to all clusters, and then to all nodes within a relevant
cluster. With some hashing and caching of cluster properties, the
process is sped up considerably.