Force register copy in conditionals' else branches. (#24)
## Summary
This PR forces else branch nodes to not get views of arrays, so no
conflict could be created with the if branch. I wanted to address just
that, but it was a little hard to get a test to fail with the limited
view-inducing slicing happening before, so I also added extra
view-inducing slicing (which is how I originally find out about the
problem). Adding the extra slicing caused some of the original
integration tests to fail, and they now pass.