Improve performance of `CombineRegisterNode` (#26)
## Summary
This PR improves the performance of `CombineRegisterNode` by skipping
multiplications when possible.

## Details and comments
About one quarter of sampling time is spent evaluating
`CombineRegisterNode`. The implementation of the node is wasteful,
because it always multiplies the combined registers, even though the
first combined register can just be set (and not multiplied against the
identity as currently happens).

Initial tests show about 7% improvement in sampling time by allowing
registers to set the values rather than multiply when possible.