Add `Node.NODE_REGISTRY` to keep track of all non-abstract nodes (#52)
## Summary

We want the ability to keep track of all non-abstract nodes so that we
can ensure serialization support and correctness for all node types.
With this registry, we will be able to ensure round-trip correctness of
all node types, and if someone adds a node in the future and forgets to
make it serializable, a test will fail.