Adding release.yml file so we can release on PyPi  (#63)
## Summary
Qiskit uses the Trusted Publisher mechanism to upload packages to PyPi.
This is adding the required workflow so we can publish.

First time doing this and it's been a while since I play around with
Github Actions, so a review from @mtreinish would be appreciated :)


<!--
Fixes/Closes/Part of: #...
-->

## Details and comments
There's no need for Environments, so I just removed it. We need to
evaluate their need at some point as it's the suggested mechanism.

Co-authored-by: Juan Gomez <juan.gomez.mosquera1@ibm.com>
Co-authored-by: Ian Hincks <ian.hincks@gmail.com>