Fix suboptimal behaviour of `GroupGatesIntoBoxes` caused by unexpected behaviour `dag.multigraph_layers` (#110)
### Summary
We always assumed that the layers returned by `dag.multigraph_layers`
are optimal for our boxing purposes. It took us a while to find an
example where this is not true, but we finally found one (more details
in #106).

This PR changes the way in which operations are grouped into boxes.
Instead of turning groups into boxes "on-the-fly" (i.e., while still
iterating over the DAG's nodes), we now turn groups into boxes after the
iteration is over. With this change, we are able to add nodes to the
earliest group that is able to welcome them, instead of being forced to
add them to the latest group, the one "currently being built."

Closes #106

---------

Co-authored-by: Ian Hincks <ian.hincks@gmail.com>