Remove NoiseOracle, replace with free TensorInterface dimensions (#164)
## Summary

We have waffled quite a bit on how noise models should enter the
Samplex.sample interface. This is the newest edition of that waffling,
though hopefully an improvement.

I do not regret the effort of making and migrating to the NoiseSource, I
think we learned a lot about it and will use those ideas down the road.
I appreciate that this PR may feel like a 180, only a week later.

## Details and comments

In attempting to document the samplex IO, and with the simultaneous
attempt to update an Executor interface with the new NoiseSource, I
found that the stateful nature of the NoiseSource within the samplex
wasn't a clean experience. Of course, there would be several ways to
keep the NoiseSource and mitigate or remove its stateful nature in the
Samplex, so in that respect perhaps this PR is heavy handed. However,
the experience of using it, which I found valuable, made me remember an
old proposal we had around the time that TensorInteface was developed
that would allow circuit dimensions to not be specified exactly, but
only consistently with other specifications. This PR is an
implementation of that idea, and introduces the concept of
`free_dimensions`. These are strings whose integer value can be deduced
when assigning data.

With this change, we can add all inputs to the Samplex in the PreSamplex
processing stage. Of particular note is that I have recombined the rates
and the Pauli terms into one input type. I should be clear that I do not
at all see this as an abandonment of the idea of the NoiseSource, but
rather the realization that I think it can and should be moved out of
scope of samplomatic, and into the scope of whatever execution agent is
needing the samples. The primary motivation that caused us to split the
rates and Pauli terms was the idea that there needed to be a strong
sense of term ordering so that users knew how to unambiguously specify
local scalings. While true, it was perhaps misguided as a user _of a
Samplex_ always has this information at hand, of course. If we extend
the notion of user to some execution framework that is invoking the
Samplex, then this is still true, and it can design its interface to
make term ordering transparent to the end user; it doesn't require the
Samplex itself to try and enforce this.

_The guide is not done yet_

---------

Co-authored-by: joshuasn <53916441+joshuasn@users.noreply.github.com>