Merge pre-propagates of fractional gates with bounded angles (#234)
## Summary
Closes #182 
Follow up to #232 

## Details and comments
This PR allows merging of `PrePropagate` nodes of static fractional
gates.

I did not add a dedicated performance test, but local tests I performed
showed about 15X improvement in sampling time of circuits with static
angles, bringing them to be about on par with parametric circuits.

I left some questions on the code, because I think the patterns are a
little fragile.