Improve boxing PM integration tests by also looking for logical equality (#239)
## Summary

Previously, we only looked for buildability. With this PR, we now also
look for logical equality.

## Details and comments

Logical equality here mainly means DAGs are isomorphic. This seems to be
the right thing to me, as opposed to, say, unitary matrix equality: we
don't want the boxing PM to be in the business of commuting stuff
around, it should simply be boxing up instructions.