Re-implement AddTerminalRightDressedBoxes (#242)
## Summary

This re-implementation adds empty right-dressed boxes instead of trying
to capture available single-qubit gates.
The idea is that AbsorbSingleQubitGates pass will be responsible.

## Details and comments

The strategy is different. Rather than trying add all boxes in their
final configuration in a single pass through the dag, we do a pass to
add all necessary boxes knowing that we're probably adding too many in
most cases, and then we go back and join boxes that can be merged into
one. This strategy leads to some pretty self-contained per-node logic,
and also makes some of the book keeping a bit neater.

---------

Co-authored-by: Samuele Ferracin <sam.ferracin@ibm.com>