Add AbsorbSingleQubitGates (#237)
## Summary

This PR changes the flow of the transpiler by treating the absorption of
single-qubit gates into boxes as a stage separate from putting
entanglers and measurements into boxes. However, it doesn't change the
behaviour of a transpiler produced by generate_boxing_pass_manager in
any case I know about.

## Details and comments

Before this PR, GroupGatesIntoBoxes and GroupMeasIntoBoxes were
responsible for both placing entanglers/meas into boxes, and also any
preceding runs of single-qubit gates.
This PR reduces their scope by leaving the single-qubit gates outside of
the boxes, and adds a new pass whose only purpose is to do this.

The reason I am interested in separating this process out is to enable
more natural interaction with how people can use barriers and boxes in a
large pass manager. If the user is programming in terms of CX, and they
want to use barriers to hint at which entangling structures should end
up in a box, it is currently tricky because the qiskit transpiler may
put single-qubit gates on both sides of the CX when converting it to,
for example , CZ. Then, neither True nor False for remove_barriers will
help them too much. If True, we might stratify against their intent. If
False, we'll end up with extra boxes with single-qubit gates.

A follow-up PR will make changes to remove_barrier behaviour in the PM
generator---I wanted to separate out those changes from these to make
the review easier.


This PR contains also changes from #242

---------

Co-authored-by: Samuele Ferracin <sam.ferracin@ibm.com>