Add `Builder.yield_from_dag` to enforce easy/hard boundary (#273)
## Summary

This PR adds the `yield_from_dag` method to the builders. While the
`PassthroughBuilder` uses the `topological_op_nodes` to iterate through
a DAG, the `LeftBoxBuilder` and `RightBoxBuilder` yield all the nodes to
ensure the convention that all easy gates between the dressing and hard
gates are implemented in the dressing and all other gates are
implemented as is. Between the easy and the hard operations, the method
yields a sentinel (right now a None) to tell the builder to switch from
one instruction mode to the other, and also to inject noise/gates. Right
now, all standard single-qubit gates are easy, but this is something we
can change and expose. Closes #59.

## Details and comments

Integration tests passing, needs more testing.

---------

Co-authored-by: Ian Hincks <ian.hincks@gmail.com>